requirements.txt
scipy
numpy
pytorch
pytorch-lightning
pyreadr
pandas
matplotlib
pickle
seaborn
collections
itertools
math
